Enhance Lidar and Rangefinder modules: add point mapping, adjust data types, and improve error logging

+++ src/protocol/global_definitions.rs
@@ -19,7 +19,7 @@ pub enum ModuleType {
 pub struct RangPoint {
     pub x: i32,
     pub y: i32,
-    pub distant: u32,
+    pub distant: u16,
 }
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
